Extract toml decoding to from_toml module
