+++ src/core/global_config.rs
@@ -6,7 +6,10 @@ use crate::{
     models::{project::Project, secondary_device::Device},
 };
 
-use super::{config::toml_try_read::TryRead, device_factories_registry::DeviceFactoryRegistry};
+use super::{
+    config::from_toml::{parse_toml_global_config, ParseTomlResult},
+    device_factories_registry::DeviceFactoryRegistry,
+};
 
 pub struct GlobalConfig {
     devices: Vec<Box<dyn Device>>,
@@ -101,20 +104,12 @@ impl GlobalConfig {
     ) -> Result<GlobalConfig, String> {
         let config_toml = config_provider.read_global_config()?;
 
-        let parsed_config = toml::from_str::<PartiallyParsedGlobalConfig>(&config_toml)
-            .map_err(|e| e.to_string())?;
-
-        let (device_errors, devices) = Self::from_toml_load_tables_of(
-            parsed_config.devices,
-            |device_table| -> Result<Box<dyn Device>, String> {
-                Self::load_device_from_toml_bloc(&device_table, &device_factories_registry)
-            },
-        );
-
-        let (project_errors, projects) = Self::from_toml_load_tables_of(
-            parsed_config.projects,
-            Self::load_project_from_toml_bloc,
-        );
+        let ParseTomlResult {
+            devices,
+            projects,
+            device_errors,
+            project_errors,
+        } = parse_toml_global_config(&config_toml, device_factories_registry)?;
 
         Self::assert_no_errors_in_config(
             &device_errors,
@@ -151,45 +146,6 @@ impl GlobalConfig {
         config_provider.write_global_config(&config_toml).unwrap();
 
         Ok(())
-    }
-
-    fn from_toml_load_tables_of<T>(
-        tables: Option<Vec<Table>>,
-        loader: impl Fn(&Table) -> Result<T, String>,
-    ) -> (Vec<String>, Vec<T>) {
-        tables
-            .unwrap_or(vec![])
-            .into_iter()
-            .map(|table| loader(&table))
-            .into_iter()
-            .partition_map(From::from)
-    }
-
-    fn load_device_from_toml_bloc(
-        device_table: &Table,
-        device_factories_registry: &DeviceFactoryRegistry,
-    ) -> Result<Box<dyn Device>, String> {
-        let name: &str = device_table.try_read("name")?;
-        let device_type: &str = device_table.try_read("type")?;
-
-        let factory = device_factories_registry
-            .get_device_factory(device_type)
-            .ok_or_else(|| "Device factory not found".to_string())?;
-
-        let device = factory.build_from_toml_table(&name, &device_table)?;
-        Ok(device)
-    }
-
-    fn load_project_from_toml_bloc(project_table: &Table) -> Result<Project, String> {
-        let name: &str = project_table.try_read("name")?;
-        let path: &str = project_table.try_read("path")?;
-        let tracking_status = project_table.try_read("tracking_status")?;
-
-        Ok(Project::new(
-            name.to_string(),
-            path.to_string(),
-            Some(tracking_status),
-        ))
     }
 
     fn assert_no_errors_in_config(
