+++ src/lib.rs
@@ -6,6 +6,7 @@ mod core {
     }
 
     mod config {
+        pub mod from_toml;
         pub mod toml_try_read;
     }
